requirements.txt
ffmpeg-python
jellyfish
python-magic
numpy
docker
